Summary:
- I began a CSV library. Since it is used only in executables app/, I updated `nix/shell-python.nix`. Configs of other projects will be OK to not include it as dependency.
- The helper function, `indexing_csv`,  for querying to CoreNLP webserver.
 - CoreNLPwebclient uses `system` in `stdlib.h` which is thread-safe.
 - I tried multi-pricessing with `exec_prog` in `similarity/corenlp_helper.cpp` but was unsucessful. Not easy to control the processes.
 - It seems combinations of pythons script, `rnn++/scripts/corenlp.py` and the `system` fuction make a bottle-neck in parallel_for of `indexing_csv`. It runs faster than single-thread, but could be improved.

Test Plan: None

Reviewers: modori, wavewave

Reviewed By: wavewave

Maniphest Tasks: T238

Differential Revision: https://uphere.phacility.com/D456

+++ rnn++/src/similarity/dep_similarity.cpp
@@ -154,7 +154,7 @@ DepSimilaritySearch::json_t DepSimilaritySearch::process_query(json_t sent_json)
     for(auto const &word : words)
         cutoff.push_back(word_cutoff.cutoff(wordUIDs[word]));
 
-    for(int i=0; i<words.size(); ++i)
+    for(size_t i=0; i<words.size(); ++i)
         fmt::print("{:<10} {:6}.uid {:6}.vocaindex : {}\n", words[i], wordUIDs[words[i]].val,
                    voca.indexmap[wordUIDs[words[i]]].val,cutoff[i]);
